refactor: move `user_matrix` property from `Actor` to `Prop3D`
